refactor(analysis-types): extract effort confidence DTO
